refactor(test): extract shared E2EPrerequisites for the e2e probe code (#222 review)
The prerequisite probes (locateSiblingPlugin / locateNode / prerequisitesMet)
were copy-pasted between AppliesEditEndToEndTests (AnglesiteBridgeTests) and
MCPClientHTTPEndToEndTests (AnglesiteCoreTests). Since those live in two
separate test targets, sharing one file isn't possible — so add a small
test-only `AnglesiteTestSupport` library target (not exposed as a product, so
the app/xcodeproj never links it) holding `enum E2EPrerequisites`, and have
both test targets depend on it.

Both tests now call E2EPrerequisites.prerequisitesMet / .locateSiblingPlugin()
/ .locateNode(); the plugin-layout and Node-discovery logic now lives in one
place. freePort/FreePortError stay local to the HTTP test (not shared).

Verified: both schemes' test bundles build, neither links
libswift_DarwinFoundation1.dylib, and both e2e tests still report `skipped`
when prerequisites are absent.
